move Member to types file
