+++ src/types.tsx
@@ -6,3 +6,25 @@ export const UserSchema = z.object({
 });
 
 export type User = z.infer<typeof UserSchema> | null;
+
+export type Member = {
+  id: number;
+  name: string;
+  surname: string;
+  province: string;
+  birthDate: string;
+  birthPlace: string;
+  email: string;
+  docType: string;
+  docId: string;
+  country: string;
+  suspendedTill: string;
+  expirationDate: string;
+  cardNumber: string;
+  isActive: boolean;
+  isDeleted: boolean;
+  status: string;
+  measure: string;
+  registrationDate: string;
+  note: string;
+};
